Use int dtype where possible

+++ ecoli/processes/listeners/rnap_data.py
@@ -55,13 +55,13 @@ class RnapData(Step):
         ports = {
             'listeners': {
                 'rnap_data': listener_schema({
-                    'rna_init_event': np.zeros(n_TUs),
+                    'rna_init_event': np.zeros(n_TUs, dtype=np.int64),
                     'active_rnap_coordinates': [-1],
                     'active_rnap_domain_indexes': [-1],
                     'active_rnap_unique_indexes': [2**31],
                     'active_rnap_on_stable_RNA_indexes': [2**31],
                     'active_rnap_n_bound_ribosomes': [-1],
-                    'rna_init_event_per_cistron': ([0.] * len(self.cistron_ids),
+                    'rna_init_event_per_cistron': ([0] * len(self.cistron_ids),
                         self.cistron_ids)
                 })
             },
